got the first seed file seeded and working
